Add launch controller logic
My current idea is to have the microcontroller sleep until inputs
change. After the "arm" switch is triggered the interrupt will fire and
the controller will poll for the "launch" button until it is pressed for
one second, or the inputs change and it goes back to sleep.

The controller will need to be power cycled after each launch, which is
an effort to prevent accidental launch triggers.